Create case line data reprocessor to harmonize
This way newer case line data with useless fields like ChartDate, and
Case_ which used to be a case creation timestamp and now is a useless “Yes” as if we need a boolean to denote that a case is a case… in the case line.

This expunges the useless age_group as well, which can be inferred
and didn’t exist in earlier case line data.

And more importantly, it adds the MD5Fingerprint field as the first
column. It’s an MD5 has of enough columns to ensure no duplicates
can be found. I had to add the two dates because there were somehow
people in the case line from the same county, age, gender, jurisdiction
and event date. Which makes sense considering the sheer amount of
cases.

A a bonus, exports four reprocessed case line data exports from March.
April, May, and June, to make diffing easier.